+++ src/Tools/DragStrokeTool.hpp
@@ -8,7 +8,7 @@
 class DragStrokeTool : public Tool {
 public:
     
-    DragStrokeTool(shared_ptr<ofFbo> _canvas, int _priority,
+    DragStrokeTool(shared_ptr<ofFbo> _canvas, float _priority,
                    ofPolyline _lineA,
                    ofPolyline _lineB,
                    ofFloatColor _cbl,
